Create README.md
I am a hands-on Sourcer with +10 years of experience working with different international markets such as Mexico, Colombia, the US, and Canada. I have worked for companies such as IBM, Accenture and AMS for clients such as Atos, VeriFone, Mass Mutual, Marsh & McLennan, Medtronic, Bristol Myers, Diebold and Bayer.

My career has guided me through specializing in IT, Digital Marketing and automation, Insurance, medical devices and Pharma roles and I can quickly adapt to and enjoy working on new profiles.

I have always been a team player with solid communication, execution and networking capabilities, able to understand business realities and establish myself as a trusted consultant and partner for stakeholders.

I believe that teamwork, the ability to understand the business, creativity and amazing candidate experience are the base of a successful recruitment strategy.